+++ CHANGES.md
@@ -3,6 +3,8 @@
 The `default-resource-path` function (or its :resource-path override in the options)
 must now return a seq of paths.
 
+Added Schema types for the options argument to `assemble-configuration`.
+
 ## 0.1.9 - 22 Nov 2015
 
 New :properties option, which specifies properties that can be substituted using